requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
transformers
torch
torchvision
torchaudio
faiss-cpu
pypdf
slowapi
limits
numpy
scikit-learn